# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MojoPedia API",
        "description": "REST API untuk sistem manajemen pengetahuan MojoPedia. Menyediakan endpoint untuk manajemen user, role, dan integrasi Keycloak SSO.",
        "contact": {
            "name": "Tim Pengembang MojoPedia",
            "email": "admin@mojopedia.go.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/knowledge": {
            "get": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Daftar repositori pengetahuan (Publik / Filterable)",
                "description": "Mendukung filter berdasarkan kata kunci (q), tipe media, kategori, label, instansi, sort, dan pagination.",
                "operationId": "getKnowledgeList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Pencarian teks judul, deskripsi, penulis",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tipe",
                        "in": "query",
                        "description": "Filter tipe: Teks, Video, Gambar, Audio",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Teks",
                                "Video",
                                "Gambar",
                                "Audio"
                            ]
                        }
                    },
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "ID kategori",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting: terbaru, terpopuler, az, za",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "terbaru"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman (default: 12)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar pengetahuan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Kirim / Unggah pengetahuan baru (Sanctum)",
                "operationId": "storeKnowledge",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "judul",
                                    "category_id",
                                    "tipe"
                                ],
                                "properties": {
                                    "judul": {
                                        "type": "string",
                                        "example": "Panduan Keamanan Siber"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tipe": {
                                        "type": "string",
                                        "enum": [
                                            "Teks",
                                            "Video",
                                            "Gambar",
                                            "Audio"
                                        ]
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Ringkasan panduan..."
                                    },
                                    "detail": {
                                        "type": "string",
                                        "example": "<p>Detail lengkap materi</p>"
                                    },
                                    "penulis": {
                                        "type": "string",
                                        "example": "Diskominfo"
                                    },
                                    "kolaborator": {
                                        "type": "string",
                                        "example": "BSSN"
                                    },
                                    "tags": {
                                        "type": "string",
                                        "example": "keamanan,sop,data"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "gambar_sampul": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengetahuan berhasil diajukan"
                    },
                    "422": {
                        "description": "Validasi form gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/knowledge/{id}": {
            "get": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Detail dokumen pengetahuan & catat view",
                "operationId": "getKnowledgeDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data pengetahuan ditemukan"
                    },
                    "404": {
                        "description": "Pengetahuan tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Knowledge"
                ],
                "summary": "Hapus dokumen pengetahuan (Pemilik atau Admin)",
                "operationId": "deleteKnowledge",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengetahuan berhasil dihapus"
                    },
                    "403": {
                        "description": "Tidak memiliki izin"
                    },
                    "404": {
                        "description": "Pengetahuan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token Sanctum Anda. Dapatkan token melalui endpoint login.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}